requirements.txt
dlt
dlt[filesystem]
dlt[parquet]
duckdb
pandas
boto3
python-dotenv
requests
PyYAML